pyproject.toml
[tool]
[tool.poetry]
name = "hamilton_example"
version = "0.1.0"
homepage = "https://github.com/ciuffredaluca/hamilton_example"
description = "Top-level package for hamilton-example."
authors = ["Luca Ciuffreda <ciuffredaluca@prometeia.com>"]
readme = "README.rst"
license =  "GPL-3.0-only"
classifiers=[
    'Development Status :: 2 - Pre-Alpha',
    'Intended Audience :: Developers',
    'License :: OSI Approved :: GNU General Public License v3 (GPLv3)',
    'Natural Language :: English',
    'Programming Language :: Python :: 3',
    'Programming Language :: Python :: 3.5',
    'Programming Language :: Python :: 3.6',
    'Programming Language :: Python :: 3.7',
    'Programming Language :: Python :: 3.8',
]
packages = [
    { include = "hamilton_example" },
    { include = "tests", format = "sdist" },
]

[tool.poetry.dependencies]
python = ">=3.10,<4"
click = "*"
sf-hamilton = ">=1.55.1"
scikit-learn = "*"

[tool.poetry.dev-dependencies]
bumpversion = "*"
coverage = "*"
flake8 = "*"
invoke = "*"
isort = "*"
pylint = "*"
pytest = ">=7.2.0"
sphinx = ">=5.3.0"
black = "*"
jupyter = "*"

[tool.poetry.scripts]
hamilton_example = 'hamilton_example.cli:main'

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[tool.isort]
profile = "black"
